Build dependency graph from Cheffile

+++ lib/remy/cookbook.rb
@@ -1,28 +1,28 @@
-require 'zlib'
-require 'archive/tar/minitar'
-require 'chef/knife/cookbook_site_download'
-
 module Remy
   class Cookbook
-    attr_reader :name, :requirement
+    attr_reader :name, :version_constraint
 
     DOWNLOAD_LOCATION = '/tmp'
 
-    def initialize name, requirement_string=">= 0"
+    def initialize name, constraint_string=">= 0.0.0"
       @name = name
-      @requirement = Gem::Requirement.create(requirement_string)
+      @version_constraint = DepSelector::VersionConstraint.new(constraint_string)
     end
 
     def download
+      return true if File.exists? download_filename
       download_command = "knife cookbook site download #{name}"
-      download_command << " #{latest_constrained_version}" unless @requirement.nil?
+      download_command << " #{latest_constrained_version}"
       `#{download_command} --file #{download_filename}`
     end
 
     def unpack
+      return true if File.exists? unpacked_cookbook_path
+      download
       fname = download_filename
       if File.exists? fname
         Archive::Tar::Minitar.unpack(Zlib::GzipReader.new(File.open(fname)), unpacked_cookbook_path)
+        true
       else
         # TODO: Raise friendly error message class
         raise "Archive hasn't been downloaded yet"
@@ -36,16 +36,17 @@ module Remy
 
     def latest_constrained_version
       versions.reverse.each do |v|
-        return v if @requirement.satisfied_by? v
+        return v if @version_constraint.include? v #TODO: Confirm this works
       end
     end
 
     def versions
-      cookbook_data['versions'].collect { |v| Gem::Version.new(v.split(/\//).last.gsub(/_/, '.')) }.sort
+      cookbook_data['versions'].collect { |v| DepSelector::Version.new(v.split(/\//).last.gsub(/_/, '.')) }.sort
     end
 
     def cookbook_data
-      @cookbook_data ||= JSON::parse(`knife cookbook site show #{@name} --format json`)
+      command = "knife cookbook site show #{@name} --format json"
+      @cookbook_data ||= JSON::parse(`#{command}`)
     end
 
     def download_filename
@@ -54,15 +55,20 @@ module Remy
 
     def unpacked_cookbook_path
       # Trimming File#extname doesn't handle the double file ext and will leave the .tar
-      File.join(download_filename.gsub(/\.tar\.gz/, ''), @name)
+      download_filename.gsub(/\.tar\.gz/, '')
     end
 
     def metadata_file
-      File.open(File.join(unpacked_cookbook_path, 'metadata.rb')).read
+      unpack
+      File.open(File.join(unpacked_cookbook_path, @name, 'metadata.rb')).read
+    end
+
+    def clean
+      FileUtils.rm_rf unpacked_cookbook_path
     end
 
     def == other
-      other.name == @name and other.requirement == @requirement
+      other.name == @name and other.version_constraint == @version_constraint
     end
   end
 end
